% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudresourcemanager_objects.R
\name{Ancestor}
\alias{Ancestor}
\title{Google Cloud Resource Manager API Objects 
The Google Cloud Resource Manager API provides methods for creating, reading, and updating project metadata.}
\usage{
Ancestor(resourceId = NULL)
}
\arguments{
\item{resourceId}{Resource id of the ancestor}
}
\value{
Ancestor object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2017-03-05 19:33:40
filename: /Users/mark/dev/R/autoGoogleAPI/googlecloudresourcemanagerv1beta1.auto/R/cloudresourcemanager_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
Ancestor Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Identifying information for a single ancestor of a project.
}
