% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudresourcemanager_functions.R
\docType{package}
\name{cloudresourcemanager_googleAuthR}
\alias{cloudresourcemanager_googleAuthR}
\alias{cloudresourcemanager_googleAuthR-package}
\title{Google Cloud Resource Manager API
The Google Cloud Resource Manager API provides methods for creating, reading, and updating project metadata.}
\description{
Auto-generated code by googleAuthR::gar_create_api_skeleton
 at 2016-09-03 23:17:29
filename: /Users/mark/dev/R/autoGoogleAPI/googlecloudresourcemanagerv1beta1.auto/R/cloudresourcemanager_functions.R
api_json: api_json
}
\details{
Authentication scopes used are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/cloud-platform.read-only
}
}

