% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudresourcemanager_objects.R
\name{FolderOperationError}
\alias{FolderOperationError}
\title{FolderOperationError Object}
\usage{
FolderOperationError(errorMessageId = NULL)
}
\arguments{
\item{errorMessageId}{The type of operation error experienced}
}
\value{
FolderOperationError object
}
\description{
FolderOperationError Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A classification of the Folder Operation error.
}
