% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudresourcemanager_objects.R
\name{GetAncestryRequest}
\alias{GetAncestryRequest}
\title{GetAncestryRequest Object}
\usage{
GetAncestryRequest()
}
\value{
GetAncestryRequest object
}
\description{
GetAncestryRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The request sent to the GetAncestry method.
}
\seealso{
Other GetAncestryRequest functions: \code{\link{projects.getAncestry}}
}

