% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudresourcemanager_objects.R
\name{GetAncestryResponse}
\alias{GetAncestryResponse}
\title{GetAncestryResponse Object}
\usage{
GetAncestryResponse(ancestor = NULL)
}
\arguments{
\item{ancestor}{Ancestors are ordered from bottom to top of the resource hierarchy}
}
\value{
GetAncestryResponse object
}
\description{
GetAncestryResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response from the GetAncestry method.
}

