% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudresourcemanager_objects.R
\name{ListOrganizationsResponse}
\alias{ListOrganizationsResponse}
\title{ListOrganizationsResponse Object}
\usage{
ListOrganizationsResponse(organizations = NULL, nextPageToken = NULL)
}
\arguments{
\item{organizations}{The list of Organizations that matched the list query, possibly paginated}

\item{nextPageToken}{A pagination token to be used to retrieve the next page of results}
}
\value{
ListOrganizationsResponse object
}
\description{
ListOrganizationsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The response returned from the `ListOrganizations` method.
}

