% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudresourcemanager_objects.R
\name{ListProjectsResponse}
\alias{ListProjectsResponse}
\title{ListProjectsResponse Object}
\usage{
ListProjectsResponse(projects = NULL, nextPageToken = NULL)
}
\arguments{
\item{projects}{The list of Projects that matched the list filter}

\item{nextPageToken}{Pagination token}
}
\value{
ListProjectsResponse object
}
\description{
ListProjectsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A page of the response received from the ListProjects method. A paginated response where more pages are available has `next_page_token` set. This token can be used in a subsequent request to retrieve the next request page.
}

