% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudresourcemanager_functions.R
\name{organizations.update}
\alias{organizations.update}
\title{Updates an Organization resource identified by the specified resource name.}
\usage{
organizations.update(Organization, name)
}
\arguments{
\item{Organization}{The \link{Organization} object to pass to this method}

\item{name}{Output Only}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/resource-manager}{Google Documentation}

Other Organization functions: \code{\link{Organization}}
}
