% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudresourcemanager_objects.R
\name{Project.labels}
\alias{Project.labels}
\title{Project.labels Object}
\usage{
Project.labels()
}
\value{
Project.labels object
}
\description{
Project.labels Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The labels associated with this Project. Label keys must be between 1 and 63 characters long and must conform to the following regular expression: \[a-z\](\[-a-z0-9\]*\[a-z0-9\])?. Label values must be between 0 and 63 characters long and must conform to the regular expression (\[a-z\](\[-a-z0-9\]*\[a-z0-9\])?)?. No more than 256 labels can be associated with a given resource. Clients should store labels in a representation such as JSON that does not depend on specific characters being disallowed. Example: 'environment' : 'dev' Read-write.
}
\seealso{
Other Project functions: \code{\link{Project}},
  \code{\link{projects.create}},
  \code{\link{projects.update}}
}

