% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudresourcemanager_objects.R
\name{Project}
\alias{Project}
\title{Project Object}
\usage{
Project(Project.labels = NULL, projectId = NULL, lifecycleState = NULL,
  projectNumber = NULL, parent = NULL, labels = NULL, createTime = NULL,
  name = NULL)
}
\arguments{
\item{Project.labels}{The \link{Project.labels} object or list of objects}

\item{projectId}{The unique, user-assigned ID of the Project}

\item{lifecycleState}{The Project lifecycle state}

\item{projectNumber}{The number uniquely identifying the project}

\item{parent}{An optional reference to a parent Resource}

\item{labels}{The labels associated with this Project}

\item{createTime}{Creation time}

\item{name}{The user-assigned display name of the Project}
}
\value{
Project object
}
\description{
Project Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A Project is a high-level Google Cloud Platform entity.  It is acontainer for ACLs, APIs, App Engine Apps, VMs, and otherGoogle Cloud Platform resources.
}
\seealso{
Other Project functions: \code{\link{Project.labels}},
  \code{\link{projects.create}},
  \code{\link{projects.update}}
}
