% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudresourcemanager_functions.R
\name{projects.create}
\alias{projects.create}
\title{Creates a Project resource.Initially, the Project resource is owned by its creator exclusively.The creator can later grant permission to others to read or update theProject.Several APIs are activated automatically for the Project, includingGoogle Cloud Storage.}
\usage{
projects.create(Project, useLegacyStack = NULL)
}
\arguments{
\item{Project}{The \link{Project} object to pass to this method}

\item{useLegacyStack}{A safety hatch to opt out of the new reliable project creation process}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/resource-manager}{Google Documentation}

Other Project functions: \code{\link{Project.labels}},
  \code{\link{Project}}, \code{\link{projects.update}}
}
