% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudresourcemanager_objects.R
\name{ResourceId}
\alias{ResourceId}
\title{ResourceId Object}
\usage{
ResourceId(type = NULL, id = NULL)
}
\arguments{
\item{type}{Required field representing the resource type this id is for}

\item{id}{Required field for the type-specific id}
}
\value{
ResourceId object
}
\description{
ResourceId Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A container to reference an id for any resource type. A `resource` in Google Cloud Platform is a generic term for something you (a developer) may want to interact with through one of our API's. Some examples are an AppEngine app, a Compute Engine instance, a Cloud SQL database, and so on.
}

